genral clean up
