+++ package.json
@@ -43,7 +43,7 @@
     "imports-loader": "0.6.5",
     "invariant": "^2.1.0",
     "isomorphic-fetch": "^2.2.1",
-    "jquery": "^3.1.0",
+    "jquery": "^3.1.1",
     "jsonp": "0.2.0",
     "key-mirror": "^1.0.1",
     "less": "^2.7.1",
