+++ subjects/HigherOrderComponents/exercise.js
@@ -9,43 +9,7 @@
 import React from 'react'
 import {render} from 'react-dom'
 
-const withMousePosition = (Component) => {
-
-  class MousePosition extends React.Component {
-    state = {
-      x: 0,
-      y: 0
-    };
-
-    constructor(props, context) {
-      super(props, context);
-
-      this.onMouseMove = this.onMouseMove.bind(this);
-    }
-
-    onMouseMove({clientX: x, clientY: y}) {
-      this.setState({x, y});
-    }
-
-
-    componentDidMount() {
-      document.addEventListener('mousemove', this.onMouseMove);
-    }
-
-    componentWillUnmount() {
-      document.removeEventListener('mousemove', this.onMouseMove);
-    }
-
-    render() {
-
-      return <Component mouse={this.state} />;
-    }
-  }
-
-  MousePosition.displayName = `MousePosition${Component.displayName || Component.name || 'Component'}`;
-
-  return MousePosition;
-};
+import withMousePosition from './withMousePosition';
 
 class App extends React.Component {
 
@@ -62,11 +26,11 @@ class App extends React.Component {
         <h1>With the mouse!</h1>
         <pre>{JSON.stringify(this.props.mouse, null, 2)}</pre>
       </div>
-    )
+    );
   }
 }
 
 const AppWithMouse = withMousePosition(App);
 
-render(<AppWithMouse/>, document.getElementById('app'));
+render(<AppWithMouse />, document.getElementById('app'));
